requirements.txt
# Augmented Intelligence Agent - Requirements

# Web Framework
streamlit>=1.28.0

# AI/ML
groq>=0.4.0

# Environment
python-dotenv>=1.0.0

# Security
bcrypt>=4.1.0

# Database (built-in sqlite3)
# No external package needed

# Utilities
pandas>=2.0.0
